moved SAML request building methods to SAMLTestRequestBuilder
